refactored List of ProjectsTab into new ProjectsListView & switched to using custom SectionIdentifier for ProjectsList
